refactor(form): rename scss files to use .module.scss
- update Select package to not use variables from Input
- remove unused checkbox variables
