+++ draft-packages/form/KaizenDraft/Form/Primitives/FieldGroup/FieldGroup.tsx
@@ -1,7 +1,7 @@
 import React, { HTMLAttributes } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "@kaizen/component-base"
-import styles from "./styles.scss"
+import styles from "./FieldGroup.module.scss"
 
 export interface FieldGroupProps
   extends OverrideClassName<HTMLAttributes<HTMLDivElement>> {
