+++ draft-packages/form/KaizenDraft/Form/Primitives/Input/Input.module.scss
@@ -3,14 +3,9 @@
 @import "~@kaizen/deprecated-component-library-helpers/styles/layout";
 @import "~@kaizen/deprecated-component-library-helpers/styles/color";
 @import "~@kaizen/component-library/styles/mixins/forms";
-@import "~@kaizen/draft-form/KaizenDraft/Form/Primitives/base";
-@import "../../decision-tokens.scss";
+@import "../../variables";
+@import "../mixins";
 
-// -----------------------------------------------
-// Form Input Primitive
-// -----------------------------------------------
-
-// Vars
 $input-height: 48px;
 $input-base-padding-horizontal: calc(#{$ca-grid} / 2);
 $input-icon-size: 1.25rem; // 20px
@@ -19,20 +14,6 @@ $input-disabled-background: $color-gray-300;
 $input-disabled-opacity: 0.3;
 $input-disabled-border-alpha: 50%;
 $input-with-icon-padding: calc(#{$input-icon-size} + calc(#{$ca-grid} * 0.75));
-
-// Helpers
-
-@mixin form-input-focus-state {
-  &:focus:not([disabled]),
-  &:hover:focus:not([disabled]),
-  &:hover:not([disabled]) {
-    @content;
-  }
-}
-
-///////////////////////////////////////////////////
-// BASE STYLES
-///////////////////////////////////////////////////
 
 .wrapper {
   position: relative;
