+++ draft-packages/form/KaizenDraft/Form/Primitives/Radio/Radio.module.scss
@@ -3,12 +3,17 @@
 @import "~@kaizen/deprecated-component-library-helpers/styles/color";
 @import "~@kaizen/component-library/styles/border";
 @import "~@kaizen/component-library/styles/mixins/forms";
-@import "../../decision-tokens.scss";
+@import "../../variables";
 
 $radio-size: 24px;
 $icon-size: 10px;
 $icon-offset: 5px;
 $focus-ring-offset: 2px;
+
+$dt-color-radio-background-color-hover: $color-gray-100;
+$dt-color-radio-disc-color-base: $color-gray-600;
+$dt-color-radio-border-color-focus: $color-blue-500;
+$dt-color-radio-border-color-focus-reversed: $color-blue-300;
 
 .radioInput {
   @include form-input-visually-hidden();
