+++ draft-packages/form/KaizenDraft/Form/Primitives/TextArea/TextArea.module.scss
@@ -1,21 +1,12 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/border";
-@import "~@kaizen/draft-form/KaizenDraft/Form/Primitives/base";
-@import "../../decision-tokens.scss";
+@import "../../variables";
+@import "../mixins";
 
 // Vars
 $input-disabled-opacity: 0.3;
 $input-disabled-border-alpha: 50%;
-
-// Helpers
-@mixin form-input-focus-state {
-  &:focus:not([disabled]),
-  &:hover:focus:not([disabled]),
-  &:hover:not([disabled]) {
-    @content;
-  }
-}
 
 .wrapper {
   position: relative;
