+++ draft-packages/form/KaizenDraft/Form/RadioField/RadioField.module.scss
@@ -1,8 +1,9 @@
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/deprecated-component-library-helpers/styles/type";
 @import "~@kaizen/deprecated-component-library-helpers/styles/color";
-@import "../Primitives/Radio/styles";
-@import "../decision-tokens.scss";
+@import "../variables";
+
+$dt-color-radio-background-color-hover: $color-gray-100;
 
 .container {
   position: relative;
