+++ draft-packages/form/KaizenDraft/Form/TextAreaField/TextAreaField.module.scss
@@ -1,6 +1,5 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/design-tokens/sass/color";
-@import "~@kaizen/draft-form/KaizenDraft/Form/Primitives/base";
 
 $input-disabled-background: $color-gray-300;
 $input-disabled-opacity: 0.3;
