+++ draft-packages/form/KaizenDraft/Form/_variables.scss
@@ -1,0 +1,12 @@
+@import "~@kaizen/design-tokens/sass/color";
+// General purpose tokens
+$dt-color-form-border-color: $color-gray-500;
+$dt-color-form-border-color-focus: $color-blue-500;
+$dt-color-form-border-color-hover: $color-gray-600;
+$dt-color-form-border-color-reversed-focus: $color-blue-300;
+$dt-color-form-text-color: $color-purple-800;
+$dt-color-form-text-color-reversed: $color-white;
+$dt-color-form-text-color-label: rgba($color-purple-800-rgb, 0.75);
+$dt-color-form-text-color-label-reversed: rgba($color-white-rgb, 0.8);
+$dt-color-form-color-icon-reversed: $color-gray-600;
+$dt-color-form-text-color-placeholder: rgba($color-purple-800-rgb, 0.7);
