--- draft-packages/form/KaizenDraft/Form/decision-tokens.scss
@@ -1,29 +1,0 @@
-@import "~@kaizen/design-tokens/sass/color";
-// General purpose tokens
-$dt-color-form-border-color: $color-gray-500;
-$dt-color-form-border-color-focus: $color-blue-500;
-$dt-color-form-border-color-hover: $color-gray-600;
-$dt-color-form-border-color-reversed-focus: $color-blue-300;
-$dt-color-form-text-color: $color-purple-800;
-$dt-color-form-text-color-reversed: $color-white;
-$dt-color-form-text-color-label: rgba($color-purple-800-rgb, 0.75);
-$dt-color-form-text-color-label-reversed: rgba($color-white-rgb, 0.8);
-$dt-color-form-color-icon-reversed: $color-gray-600;
-$dt-color-form-text-color-placeholder: rgba($color-purple-800-rgb, 0.7);
-
-// component specific tokens
-$dt-color-checkbox-background-color-checked: $color-gray-500;
-$dt-color-checkbox-background-color-checked-focus: $color-gray-600;
-$dt-color-checkbox-background-color-unchecked-focus: $color-gray-100;
-$dt-color-checkbox-background-color-hover: $color-gray-600;
-$dt-color-checkbox-background-color-unchecked-hover: $color-gray-100;
-$dt-color-checkbox-border-color-checked-focus: $color-gray-600;
-$dt-color-checkbox-border-color-focus: $color-gray-600;
-$dt-color-checkbox-border-color-hover: $color-gray-600;
-$dt-color-radio-disc-color-base: $color-gray-600;
-$dt-color-radio-background-color-hover: $color-gray-100;
-$dt-color-radio-border-color-focus: $color-blue-500;
-$dt-color-radio-border-color-focus-reversed: $color-blue-300;
-$dt-color-toggle-background-color: $color-gray-300;
-$dt-color-toggle-background-color-hover: $color-purple-500;
-$dt-color-toggle-background-color-hover: rgba($color-gray-500-rgb, 0.4);
