+++ draft-packages/select/KaizenDraft/Select/styles.react.scss
@@ -2,10 +2,13 @@
 @import "~@kaizen/design-tokens/sass/shadow";
 @import "~@kaizen/design-tokens/sass/border";
 @import "~@kaizen/design-tokens/sass/color";
-@import "~@kaizen/draft-form/KaizenDraft/Form/Primitives/Input/styles.scss";
 @import "~@kaizen/deprecated-component-library-helpers/styles/color";
 @import "~@kaizen/component-library/styles/border";
 @import "~@kaizen/deprecated-component-library-helpers/styles/type";
+
+// Should match the values for the Input component
+$input-height: 48px;
+$input-placeholder-opacity: 0.5;
 
 $focus-border-color: $color-blue-500;
